Miscellaneous patches (#166)
* Miscellaneous patches

- compile-time-perf
- updated codecov
- doxygen updates
- updated ex_cxx_overhead
- updated ex-gotcha
- updated ex-gpu-roofline
- updated ex-statistics
- reduced number of intermediate libraries
- updated pthread_gotcha
- fixed bundler rekey support
- removed const ref to fundamental types
- simplified fake half2
- source location updates
- manager get_is_main_thread
- migrated runtime enabled checks into generic_operator
- preliminary support for positional arguments in argparse
- fixed signal handler to avoid debugging
- simplified type_mutex
- fixed bundle copy-operator
- fixed lightweight tuple +=, -=
- updated timemory-avail to fit into window
- fixed loading mpi and nccl libraries

* Fixes to compiler-instrumentation and gotcha replacement

* Minor updates

* PTL and compiler-instrument fixes

- disabled pthread-gotcha by default in compiler instrumentation
- removed main-gotcha in compiler instrumentation
- global lock in compiler instrumentation
- miscellaneous shadow and double-promotion fixes

* Fixed ex_gotcha_replacement

- beware of -ffinite-math-only w/ math gotcha
- timemory::timemory-compile-options does not automatically provide timemory::timemory-compile-extra in install tree
- disabled unity builds for examples

* Removed main_gotcha reference in compiler instrumentation

* Remove debug info from compiler instrumentation

+++ source/tests/cuda_tests.cpp
@@ -160,8 +160,8 @@ TEST_F(cuda_tests, saxpy)
     float sumError = 0.0;
     for(int64_t i = 0; i < N; i++)
     {
-        maxError = std::max<float>(maxError, std::abs(y[i] - 2.0));
-        sumError += (y[i] > 2.0) ? (y[i] - 2.0) : (2.0 - y[i]);
+        maxError = std::max<float>(maxError, std::abs(y[i] - 2.0f));
+        sumError += (y[i] > 2.0f) ? (y[i] - 2.0f) : (2.0f - y[i]);
     }
 
     tim::device::cpu::free(x);
@@ -176,12 +176,12 @@ TEST_F(cuda_tests, saxpy)
 #endif
     auto rc = *bw.get<wall_clock>();
 
-    printf("Max error: %8.4e\n", maxError);
-    printf("Sum error: %8.4e\n", sumError);
-    printf("Total amount of data (GB): %f\n", data_size);
-    printf("Effective Bandwidth (GB/s): %f\n", data_size / ce.get());
-    printf("Kernel Runtime (sec): %16.12e\n", ce.get());
-    printf("Wall-clock time (sec): %16.12e\n", rc.get());
+    printf("Max error: %8.4e\n", (double) maxError);
+    printf("Sum error: %8.4e\n", (double) sumError);
+    printf("Total amount of data (GB): %f\n", (double) data_size);
+    printf("Effective Bandwidth (GB/s): %f\n", (double) (data_size / ce.get()));
+    printf("Kernel Runtime (sec): %16.12e\n", (double) ce.get());
+    printf("Wall-clock time (sec): %16.12e\n", (double) rc.get());
     std::cout << details::get_test_name() << " cuda event: " << ce << std::endl;
     std::cout << details::get_test_name() << " real clock: " << rc << std::endl;
     std::cout << tot << std::endl;
@@ -277,8 +277,8 @@ TEST_F(cuda_tests, saxpy_streams)
     float sumError = 0.0;
     for(int64_t i = 0; i < N; i++)
     {
-        maxError = std::max<float>(maxError, std::abs(y[i] - 2.0));
-        sumError += (y[i] > 2.0) ? (y[i] - 2.0) : (2.0 - y[i]);
+        maxError = std::max<float>(maxError, std::abs(y[i] - 2.0f));
+        sumError += (y[i] > 2.0f) ? (y[i] - 2.0f) : (2.0f - y[i]);
     }
 
     tim::device::cpu::free(x);
@@ -293,12 +293,12 @@ TEST_F(cuda_tests, saxpy_streams)
 #endif
     auto rc = *bw.get<wall_clock>();
 
-    printf("Max error: %8.4e\n", maxError);
-    printf("Sum error: %8.4e\n", sumError);
-    printf("Total amount of data (GB): %f\n", data_size);
-    printf("Effective Bandwidth (GB/s): %f\n", data_size / ce.get());
-    printf("Kernel Runtime (sec): %16.12e\n", ce.get());
-    printf("Wall-clock time (sec): %16.12e\n", rc.get());
+    printf("Max error: %8.4e\n", (double) maxError);
+    printf("Sum error: %8.4e\n", (double) sumError);
+    printf("Total amount of data (GB): %f\n", (double) data_size);
+    printf("Effective Bandwidth (GB/s): %f\n", (double) (data_size / ce.get()));
+    printf("Kernel Runtime (sec): %16.12e\n", (double) ce.get());
+    printf("Wall-clock time (sec): %16.12e\n", (double) rc.get());
     std::cout << details::get_test_name() << " cuda event: " << ce << std::endl;
     std::cout << details::get_test_name() << " real clock: " << rc << std::endl;
     std::cout << tot << std::endl;
